SMTP: Do not delete files in test mode

+++ squirt/smtplayer.py
@@ -97,7 +97,7 @@ def smtp_send(script):
                            filename=filename)
             outer.attach(msg)
 
-            if script.get('delete')==True:
+            if script.get('delete')==True and script.get('test')==False:
                 os.remove(path)
 
     # And then send the message
